backend/requirements.txt
alembic==1.12.0
annotated-types==0.7.0
anyio==3.7.1
black==23.10.0
click==8.2.0
exceptiongroup==1.3.0
fastapi==0.104.0
greenlet==3.2.2
h11==0.16.0
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.9
pydantic==2.11.4
pydantic_core==2.33.2
pytest==8.3.5
python-dotenv==1.0.0
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.22
starlette==0.27.0
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.23.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
pydantic-settings==2.1.0